Fix file arguments handling when using --chdir
