+++ cmd/inspect.go
@@ -2,6 +2,8 @@ package cmd
 
 import (
 	"fmt"
+	"os"
+	"path/filepath"
 	"strings"
 
 	"github.com/hashicorp/hcl/v2"
@@ -14,18 +16,9 @@ import (
 	"google.golang.org/grpc/status"
 )
 
-func (cli *CLI) inspect(opts Options, targetDir string, filterFiles []string) int {
+func (cli *CLI) inspect(opts Options, args []string) int {
 	// Respect the "--format" flag until a config is loaded
 	cli.formatter.Format = opts.Format
-
-	if opts.Chdir != "" && targetDir != "." {
-		cli.formatter.Print(tflint.Issues{}, fmt.Errorf("Cannot use --chdir and directory argument at the same time"), map[string][]byte{})
-		return ExitCodeError
-	}
-	if opts.Recursive && (targetDir != "." || len(filterFiles) > 0) {
-		cli.formatter.Print(tflint.Issues{}, fmt.Errorf("Cannot use --recursive and arguments at the same time"), map[string][]byte{})
-		return ExitCodeError
-	}
 
 	workingDirs, err := findWorkingDirs(opts)
 	if err != nil {
@@ -37,6 +30,23 @@ func (cli *CLI) inspect(opts Options, targetDir string, filterFiles []string) in
 
 	for _, wd := range workingDirs {
 		err := cli.withinChangedDir(wd, func() error {
+			// Parse directory/file arguments after changing the working directory
+			targetDir, filterFiles, err := processArgs(args[1:])
+			if err != nil {
+				return fmt.Errorf("Failed to parse CLI arguments; %w", err)
+			}
+
+			if opts.Chdir != "" && targetDir != "." {
+				return fmt.Errorf("Cannot use --chdir and directory argument at the same time")
+			}
+			if opts.Recursive && (targetDir != "." || len(filterFiles) > 0) {
+				return fmt.Errorf("Cannot use --recursive and arguments at the same time")
+			}
+
+			// Join with the working directory to create the fullpath
+			for i, file := range filterFiles {
+				filterFiles[i] = filepath.Join(wd, file)
+			}
 			moduleIssues, err := cli.inspectModule(opts, targetDir, filterFiles)
 			if err != nil {
 				return err
@@ -71,6 +81,49 @@ func (cli *CLI) inspect(opts Options, targetDir string, filterFiles []string) in
 	}
 
 	return ExitCodeOK
+}
+
+func processArgs(args []string) (string, []string, error) {
+	if len(args) == 0 {
+		return ".", []string{}, nil
+	}
+
+	var dir string
+	filterFiles := []string{}
+
+	for _, file := range args {
+		fileInfo, err := os.Stat(file)
+		if err != nil {
+			if os.IsNotExist(err) {
+				return dir, filterFiles, fmt.Errorf("Failed to load `%s`: File not found", file)
+			}
+			return dir, filterFiles, fmt.Errorf("Failed to load `%s`: %s", file, err)
+		}
+
+		if fileInfo.IsDir() {
+			dir = file
+			if len(args) != 1 {
+				return dir, filterFiles, fmt.Errorf("Failed to load `%s`: Multiple arguments are not allowed when passing a directory", file)
+			}
+			return dir, filterFiles, nil
+		}
+
+		if !strings.HasSuffix(file, ".tf") && !strings.HasSuffix(file, ".tf.json") {
+			return dir, filterFiles, fmt.Errorf("Failed to load `%s`: File is not a target of Terraform", file)
+		}
+
+		fileDir := filepath.Dir(file)
+		if dir == "" {
+			dir = fileDir
+			filterFiles = append(filterFiles, file)
+		} else if fileDir == dir {
+			filterFiles = append(filterFiles, file)
+		} else {
+			return dir, filterFiles, fmt.Errorf("Failed to load `%s`: Multiple files in different directories are not allowed", file)
+		}
+	}
+
+	return dir, filterFiles, nil
 }
 
 func (cli *CLI) inspectModule(opts Options, dir string, filterFiles []string) (tflint.Issues, error) {
