+++ integrationtest/cli/chdir/main.tf
@@ -1,0 +1,3 @@
+resource "aws_instance" "main" {
+  instance_type = "t2.micro"
+}
