+++ integrationtest/cli/chdir/subdir/.tflint.hcl
@@ -1,0 +1,3 @@
+plugin "testing" {
+  enabled = true
+}
