+++ integrationtest/cli/chdir/subdir/main.tf
@@ -1,0 +1,3 @@
+resource "aws_instance" "main" {
+  instance_type = "m5.2xlarge"
+}
