+++ integrationtest/cli/cli_test.go
@@ -276,22 +276,43 @@ func TestIntegration(t *testing.T) {
 		},
 		{
 			name:    "--chdir",
-			command: "./tflint --chdir subdir",
-			dir:     "multiple_files",
+			command: "./tflint --chdir=subdir",
+			dir:     "chdir",
+			status:  cmd.ExitCodeIssuesFound,
+			stdout:  fmt.Sprintf("%s (aws_instance_example_type)", color.New(color.Bold).Sprint("instance type is m5.2xlarge")),
+		},
+		{
+			name:    "--chdir and file argument",
+			command: "./tflint --chdir=subdir main.tf",
+			dir:     "chdir",
 			status:  cmd.ExitCodeIssuesFound,
 			stdout:  fmt.Sprintf("%s (aws_instance_example_type)", color.New(color.Bold).Sprint("instance type is m5.2xlarge")),
 		},
 		{
 			name:    "--chdir and directory argument",
-			command: "./tflint --chdir subdir ../",
-			dir:     "multiple_files",
+			command: "./tflint --chdir=subdir ../",
+			dir:     "chdir",
 			status:  cmd.ExitCodeError,
 			stderr:  "Cannot use --chdir and directory argument at the same time",
 		},
 		{
-			name:    "--recursive and arguments",
+			name:    "--chdir and file under the directory argument",
+			command: fmt.Sprintf("./tflint --chdir=subdir %s", filepath.Join("nested", "main.tf")),
+			dir:     "chdir",
+			status:  cmd.ExitCodeError,
+			stderr:  "Cannot use --chdir and directory argument at the same time",
+		},
+		{
+			name:    "--recursive and file argument",
+			command: "./tflint --recursive main.tf",
+			dir:     "chdir",
+			status:  cmd.ExitCodeError,
+			stderr:  "Cannot use --recursive and arguments at the same time",
+		},
+		{
+			name:    "--recursive and directory argument",
 			command: "./tflint --recursive subdir",
-			dir:     "multiple_files",
+			dir:     "chdir",
 			status:  cmd.ExitCodeError,
 			stderr:  "Cannot use --recursive and arguments at the same time",
 		},
